+++ src/main/java/com/sparta/myblogserver/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.myblogserver.domain.repository;
+package com.sparta.myblogserver.repository;
 
-import com.sparta.myblogserver.domain.entity.post.Post;
+import com.sparta.myblogserver.entity.post.Post;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
